Move test classes into test directory
We previously put these classes in the sources directory because they
needed to be available to both the unit tests and the benchmarking code.
Since we currently don't have benchmarking code, we can safely move them
for now. If we add benchmarking in the future, we should be able to add
both library sources and test sources to the build target.