Add unit test coverage to RandomCutForest
Add missing unit tests to cover extrapolation, near neighbors, samplers
full, and total updates.

This commit also adds new checks to extrapolateBasic to validate the
shingleIndex value. In a cyclic shingle, the shingle index refers to the
position in the shingle where the next "raw" input point should be
written. In terms of extrapolateBasic arguments, the shingle index
should be a value between 0 (inclusive) and dimensions / blockSize
(exclusive).

Closes #10