Remove compact node manager class
In a previous revision, we introducted the CompactNodeManager class to
unify the handling of leaf nodes and internal nodes. The initial
implementation was a leaky abstraction, because the tree class was still
responsible for managing some node properties, like bounding boxes.
The change moves node management back into the tree class.