[Rust] Create a Visitor trait to abstract out scoring algorithms (#217)
* [Rust] Create a Visitor trait to abstract out scoring algorithms

Most scoring algorithms make use of a visitor design pattern. The Visitor trait
makes it easy to perform node traversals with new algorithms. We also add a
Tree::traverse() method that makes use of any type that implements Visitor.
This addition makes the implementation of anomaly_score() much cleaner.

* [Rust] Check for Visitor impl at compile-time

* Create a separate Tree::traverse function

The way node iteration is implemented requires collecting and reversing
the nodes before applying a visitor. We create an independent traverse
function specifically for visitors. The existing iter() function is
left as-is for testing purposes.

Closes #219

* [Rust] Rename algorithm submodule to visitor

* [Rust] Fix docstrings